Moved ErrorMessage tests into separate file
