+++ static/js/containers/DashboardPage_test.js
@@ -2,30 +2,16 @@
 import '../global_init';
 
 import { assert } from 'chai';
-import _ from 'lodash';
 
-import {
-  REQUEST_DASHBOARD,
-  RECEIVE_DASHBOARD_FAILURE,
-  RECEIVE_GET_USER_PROFILE_SUCCESS,
-} from '../actions';
-import {
-  DASHBOARD_RESPONSE,
-  ERROR_RESPONSE,
-} from '../constants';
 import IntegrationTestHelper from '../util/integration_test_helper';
+import { REQUEST_DASHBOARD } from '../actions';
 
 describe('DashboardPage', () => {
   let renderComponent, helper;
-  let dashboardErrorActions;
 
   beforeEach(() => {
     helper = new IntegrationTestHelper();
     renderComponent = helper.renderComponent.bind(helper);
-    dashboardErrorActions = [
-      RECEIVE_DASHBOARD_FAILURE,
-      RECEIVE_GET_USER_PROFILE_SUCCESS
-    ];
   });
 
   afterEach(() => {
@@ -38,59 +24,6 @@ describe('DashboardPage', () => {
       helper.store.dispatch({ type: REQUEST_DASHBOARD });
 
       assert(div.querySelector(".spinner"), "Unable to find spinner");
-    });
-  });
-
-  describe('dashboard errors', () => {
-    let errorString = `Sorry, we were unable to load the data necessary
-      to process your request. Please reload the page.`;
-    errorString = errorString.replace(/\s\s+/g, ' ');
-
-    it('error from the backend triggers error message in dashboard', () => {
-      helper.dashboardStub.returns(Promise.reject(ERROR_RESPONSE));
-
-      return renderComponent("/dashboard", dashboardErrorActions, false).then(([, div]) => {
-        let message = div.getElementsByClassName('alert-message')[0];
-        assert(message.textContent.indexOf(errorString) > -1);
-        assert(message.textContent.indexOf(ERROR_RESPONSE.error_code) > -1);
-        assert(message.textContent.indexOf("Additional info:") > -1);
-        assert(message.textContent.indexOf(ERROR_RESPONSE.user_message) > -1);
-      });
-    });
-
-    it('the error from the backend does not need to be complete', () => {
-      let response = _.cloneDeep(ERROR_RESPONSE);
-      delete response.user_message;
-      helper.dashboardStub.returns(Promise.reject(response));
-
-      return renderComponent("/dashboard", dashboardErrorActions, false).then(([, div]) => {
-        let message = div.getElementsByClassName('alert-message')[0];
-        assert(message.textContent.indexOf(errorString) > -1);
-        assert(message.textContent.indexOf(ERROR_RESPONSE.error_code) > -1);
-        assert.equal(message.textContent.indexOf("Additional info:"), -1);
-        assert.equal(message.textContent.indexOf(ERROR_RESPONSE.user_message), -1);
-      });
-    });
-
-    it('the error from the backend does not need to exist at all as long as there is an http error', () => {
-      helper.dashboardStub.returns(Promise.reject({}));
-
-      return renderComponent("/dashboard", dashboardErrorActions, false).then(([, div]) => {
-        let message = div.getElementsByClassName('alert-message')[0];
-        assert(message.textContent.indexOf(errorString) > -1);
-        assert.equal(message.textContent.indexOf(ERROR_RESPONSE.error_code), -1);
-        assert.equal(message.textContent.indexOf("Additional info:"), -1);
-        assert.equal(message.textContent.indexOf(ERROR_RESPONSE.user_message), -1);
-      });
-    });
-
-    it('a regular response does not show the error', () => {
-      helper.dashboardStub.returns(Promise.resolve(DASHBOARD_RESPONSE));
-
-      return renderComponent("/dashboard").then(([, div]) => {
-        let message = div.getElementsByClassName('alert-message')[0];
-        assert.equal(message, undefined);
-      });
     });
   });
 });
